refactor: move parseGitHubIssueUrl to dependency-free lib/github/url.ts
Resolves client/server boundary violation in request-task-modal.tsx.
The modal is a client component but was importing from repo-profile.ts
which pulls in GitHub API helpers, process.env.GITHUB_TOKEN, and Buffer.
The pure URL-parsing function now lives in lib/github/url.ts with no
server-only dependencies. repo-profile.ts re-exports it for backward
compatibility; all other import sites updated to the new path.
