+++ lib/github/repo-profile.ts
@@ -6,6 +6,7 @@ import {
   fetchFileContent,
 } from './client'
 import { detectStack } from './stack-detection'
+export { parseGitHubIssueUrl } from './url'
 
 // ---------------------------------------------------------------------------
 // buildRepoProfile
@@ -80,28 +81,3 @@ export async function buildRepoProfile(owner: string, repo: string): Promise<Rep
   }
 }
 
-// ---------------------------------------------------------------------------
-// parseGitHubIssueUrl
-// ---------------------------------------------------------------------------
-
-/**
- * Parse a GitHub issue URL and return the owner, repo name, and issue number.
- * Returns `null` when the URL does not match the expected format.
- *
- * @example
- * parseGitHubIssueUrl('https://github.com/vercel/next.js/issues/12345')
- * // => { owner: 'vercel', repo: 'next.js', issueNumber: 12345 }
- */
-export function parseGitHubIssueUrl(url: string): {
-  owner: string
-  repo: string
-  issueNumber: number
-} | null {
-  const match = url.match(/github\.com\/([\w.-]+)\/([\w.-]+)\/issues\/(\d+)/)
-  if (!match) return null
-  return {
-    owner: match[1],
-    repo: match[2],
-    issueNumber: parseInt(match[3], 10),
-  }
-}
